requirementsv2.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.1
django-bootstrap3==25.1
django-bootstrap4==25.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-mssql-backend==2.8.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib==19.24.8
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
future==1.0.0
html5lib==1.1
idna==3.10
lxml==5.4.0
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.2.5
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.30.2
psycopg2==2.9.10
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pyodbc==5.2.0
pypdf==5.5.0
PyPDF2==3.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
xhtml2pdf==0.2.17
XlsxWriter==3.2.3
